requirements.txt
pylint
pytest
selenium
geopy
p_tqdm
pandas
numpy
python-dotenv